move hook-utils
