+++ frontend/packages/console-app/src/components/cloud-shell/__tests__/useActivityTick.spec.ts
@@ -1,6 +1,6 @@
 import useActivityTick from '../useActivityTick';
 import * as cloudShellUtils from '../cloud-shell-utils';
-import { testHook } from '@console/shared/src/test-utils/hooks-utils';
+import { testHook } from '../../../../../../__tests__/utils/hooks-utils';
 
 describe('useActivityTick', () => {
   beforeEach(() => {
