+++ frontend/packages/console-shared/src/hooks/__tests__/select-list.spec.ts
@@ -1,5 +1,5 @@
 import { act } from 'react-dom/test-utils';
-import { testHook } from '../../test-utils/hooks-utils';
+import { testHook } from '../../../../../__tests__/utils/hooks-utils';
 import { useSelectList } from '../select-list';
 import { data, visibleRows, onRowSelected } from '../__mocks__/select-list-data';
