+++ frontend/packages/pipelines-plugin/src/components/pipelines/__tests__/hooks.spec.ts
@@ -1,5 +1,5 @@
 import { useK8sWatchResource } from '@console/internal/components/utils/k8s-watch-hook';
-import { testHook } from '@console/shared/src/test-utils/hooks-utils';
+import { testHook } from '../../../../../../__tests__/utils/hooks-utils';
 import { pipelineTestData, PipelineExampleNames } from '../../../test-data/pipeline-data';
 import { PipelineRun } from '../../../utils/pipeline-augment';
 import * as hooks from '../hooks';
